sched/fair: Decide 'eenv->sg_cap' ahead energy computation
eenv::sg_cap is used to indicate which sched_group CPUs are bound to
the same clock domain, as result the final capacity needs to consider
all CPUs in eenv::sg_cap.  Now eenv::sg_cap is set in the middle
of flow for every sched_group energy calculation, but this is not
necessary and we can prepare its value ahead energy computation.

This patch moves eenv::sg_cap setting from sched_group_energy() to
function compute_task_energy(), so can prepare it ahead energy
computation and this can avoid to set it repeatedly.

Change-Id: I9e231fdc26bbf61aa09894c9ea0532b3aa2f3f6b
Signed-off-by: Leo Yan <leo.yan@linaro.org>
Signed-off-by: mydongistiny <jaysonedson@gmail.com>